moved login popup outside of the contact information form
- this was necessary as submission of the login
form cause the outer form to submit as well,
resulting in an unwanted behavior
